Breakdown the Composite Loot index.js file into smaller files
